Improve [Backend] [Middleware] [Authentication] OAuth2 Web Version for better extensibility and maintainability
- [+] refactor(authentication): move user info retrieval logic to a separate function and remove unused buffer code
- [+] feat(authentication): add getUserInfo function to handle user information retrieval from OAuth2 provider
